feat!: remove per-tenant metrics export
This should never be part of Supavisor in the first place. This should
be handled at different level instead. In the same vain it removes
fetching metrics cluster-wide, instead operator is encouraged to gather
metrics from each node independently. This is less error prone and
provides better way to monitor individual nodes health.